Moved Window Factory into the controller.
